+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Remove prometheus remote write agent configuration when mimir is enabled.
+- Remove unnecessary prometheus control-plane affinity.
 
 ## [4.75.0] - 2024-05-13